Restructure svelte files
move svelte components to their own directory
