--- src/routes/message.svelte
@@ -1,22 +1,0 @@
-<div
-class="w-1/6 lg:w-1/12 fixed top-0 left-0" id="message">
-    <svg
-        id="message"
-        class="relative w-1/2"
-        viewBox="0 0 512 512"
-        xmlns="http://www.w3.org/2000/svg"
-        ><path
-            d="M467 61H45C20.218 61 0 81.196 0 106v300c0 24.72 20.128 45 45 45h422c24.72 0 45-20.128 45-45V106c0-24.72-20.128-45-45-45zm-6.214 30L256.954 294.833 51.359 91h409.427zM30 399.788V112.069l144.479 143.24L30 399.788zM51.213 421l144.57-144.57 50.657 50.222c5.864 5.814 15.327 5.795 21.167-.046L317 277.213 460.787 421H51.213zM482 399.787L338.213 256 482 112.212v287.575z"
-        /><path
-            d="M154.661 192.852L52.373 91.452l101.814-.161a78521.3 78521.3 0 01203.707 0l101.895.16-101.42 101.4-101.42 101.401zM410.263 327.356l-71.356-71.358 71.516-71.513 71.515-71.514v142.872c0 78.58-.071 142.871-.16 142.871-.087 0-32.269-32.11-71.515-71.358z"
-            id="path1"
-            fill="#9fdcff"
-            fill-rule="evenodd"
-        /><path
-            d="M123.882 348.88l71.984-71.984 25.795 25.72c16.74 16.692 26.73 26.187 28.46 27.05 3.655 1.825 9.992 1.828 13.657.006 1.736-.863 11.485-10.123 27.982-26.58l25.317-25.256 71.512 71.515 71.513 71.514H51.897zM30.062 256.104V113.179l2.373 1.95c1.305 1.072 33.646 33.064 71.868 71.093l69.495 69.145-71.868 71.831-71.868 71.83z"
-            id="path2"
-            fill="#9fdcff"
-            fill-rule="evenodd"
-        /></svg
-    >
-</div>
